+++ PersonApi.Tests/V1/Gateways/DynamoDbGatewayTests.cs
@@ -29,38 +29,45 @@ namespace PersonApi.Tests.V1.Gateways
         [Test]
         public async Task GetPersonByIdReturnsNullIfEntityDoesntExist()
         {
+            // Act
             var response = await _classUnderTest.GetPersonByIdAsync(Guid.NewGuid()).ConfigureAwait(false);
 
+            // Assert
             response.Should().BeNull();
         }
 
         [Test]
         public async Task GetPersonByIdReturnsThePersonIfItExists()
         {
+            // Arrange
             var entity = _fixture.Create<Person>();
             var dbEntity = entity.ToDatabase();
-
             var dbIdUsed = entity.Id;
+
             _dynamoDb.Setup(x => x.LoadAsync<PersonDbEntity>(dbIdUsed, default))
                      .ReturnsAsync(dbEntity);
 
+            // Act
             var response = await _classUnderTest.GetPersonByIdAsync(entity.Id).ConfigureAwait(false);
 
+            // Assert
             _dynamoDb.Verify(x => x.LoadAsync<PersonDbEntity>(dbIdUsed, default), Times.Once);
-
             entity.Id.Should().Be(response.Id);
         }
 
         [Test]
         public void GetPersonByIdExceptionThrow()
         {
+            // Arrange
             var id = Guid.NewGuid();
             var exception = new ApplicationException("Test exception");
             _dynamoDb.Setup(x => x.LoadAsync<PersonDbEntity>(id, default))
                      .ThrowsAsync(exception);
 
+            // Act
             Func<Task<Person>> func = async () => await _classUnderTest.GetPersonByIdAsync(id).ConfigureAwait(false);
 
+            // Assert
             func.Should().Throw<ApplicationException>().WithMessage(exception.Message);
             _dynamoDb.Verify(x => x.LoadAsync<PersonDbEntity>(id, default), Times.Once);
         }
